Change board and extramen variables to longs
git-svn-id: https://svn.so-much-stuff.com/svn/trunk/cvs/trunk@1151 93276802-5317-4dee-a4ef-7ae40ebda4b9

+++ games.d/centipede.d/extern.c
@@ -19,8 +19,8 @@ COORD guy = {22,28};            /* your coordinates */
 COORD shot;                     /* the shot's coordinates */
 int fired = 0;                  /* a shot has been fired */
 long score = 0;
-int board = 1;
-int extramen = 3;
+long board = 1;
+long extramen = 3;
 long nextman = FREEMAN;
 int finished;                   /* time since board ended */
 int breeding;                   /* are they breeding */
